Complete v4 progression and discovery

+++ src/main/java/atomorphosis/cannedcuisine/client/CompactFoodTooltip.java
@@ -2,5 +2,5 @@ package atomorphosis.cannedcuisine.client;
 
 import net.minecraft.world.inventory.tooltip.TooltipComponent;
 
-public record CompactFoodTooltip(int nutrition, float saturation) implements TooltipComponent {
+public record CompactFoodTooltip(double nutrition, double saturation) implements TooltipComponent {
 }
